requirements.txt
# install modules with: pip install -r requirements.txt
bottle
cherrypy
requests
beautifulsoup4
RPi.GPIO
pychrome
